Extract `language_model` and `completion` crates
