+++ crates/collab/Cargo.toml
@@ -30,6 +30,7 @@ chrono.workspace = true
 clock.workspace = true
 clickhouse.workspace = true
 collections.workspace = true
+completion.workspace = true
 dashmap = "5.4"
 envy = "0.4.2"
 futures.workspace = true
