+++ crates/collab/src/tests/test_server.rs
@@ -294,7 +294,7 @@ impl TestServer {
             menu::init();
             dev_server_projects::init(client.clone(), cx);
             settings::KeymapFile::load_asset(os_keymap, cx).unwrap();
-            assistant::FakeCompletionProvider::setup_test(cx);
+            completion::FakeCompletionProvider::setup_test(cx);
             assistant::context_store::init(&client);
         });
