+++ crates/completion/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "assistant"
+name = "completion"
 version = "0.1.0"
 edition = "2021"
 publish = false
@@ -9,7 +9,7 @@ license = "GPL-3.0-or-later"
 workspace = true
 
 [lib]
-path = "src/assistant.rs"
+path = "src/completion.rs"
 doctest = false
 
 [features]
@@ -23,67 +23,33 @@ test-support = [
 [dependencies]
 anthropic = { workspace = true, features = ["schemars"] }
 anyhow.workspace = true
-assets.workspace = true
-assistant_slash_command.workspace = true
-async-watch.workspace = true
-breadcrumbs.workspace = true
-cargo_toml.workspace = true
-chrono.workspace = true
 client.workspace = true
-clock.workspace = true
 collections.workspace = true
-command_palette_hooks.workspace = true
-completion.workspace = true
 editor.workspace = true
-feature_flags.workspace = true
-fs.workspace = true
 futures.workspace = true
-fuzzy.workspace = true
 gpui.workspace = true
-heed.workspace = true
-html_to_markdown.workspace = true
 http.workspace = true
-indexed_docs.workspace = true
-indoc.workspace = true
-language.workspace = true
 language_model.workspace = true
 log.workspace = true
 menu.workspace = true
-multi_buffer.workspace = true
 ollama = { workspace = true, features = ["schemars"] }
 open_ai = { workspace = true, features = ["schemars"] }
-ordered-float.workspace = true
 parking_lot.workspace = true
-paths.workspace = true
-project.workspace = true
-regex.workspace = true
-rope.workspace = true
-schemars.workspace = true
-search.workspace = true
-semantic_index.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 settings.workspace = true
-similar.workspace = true
 smol.workspace = true
-telemetry_events.workspace = true
-terminal.workspace = true
-terminal_view.workspace = true
+strum.workspace = true
 theme.workspace = true
-toml.workspace = true
+tiktoken-rs.workspace = true
 ui.workspace = true
 util.workspace = true
-uuid.workspace = true
-workspace.workspace = true
-picker.workspace = true
-roxmltree = "0.20.0"
 
 [dev-dependencies]
 ctor.workspace = true
 editor = { workspace = true, features = ["test-support"] }
 env_logger.workspace = true
 language = { workspace = true, features = ["test-support"] }
-log.workspace = true
 project = { workspace = true, features = ["test-support"] }
 rand.workspace = true
 text = { workspace = true, features = ["test-support"] }
