+++ crates/completion/src/anthropic.rs
@@ -1,15 +1,12 @@
-use crate::assistant_settings::CloudModel;
-use crate::assistant_settings::{AssistantProvider, AssistantSettings};
-use crate::LanguageModelCompletionProvider;
-use crate::{
-    assistant_settings::OpenAiModel, CompletionProvider, LanguageModel, LanguageModelRequest, Role,
-};
+use crate::{count_open_ai_tokens, LanguageModelCompletionProvider};
+use crate::{CompletionProvider, LanguageModel, LanguageModelRequest};
+use anthropic::{stream_completion, Model as AnthropicModel, Request, RequestMessage};
 use anyhow::{anyhow, Result};
 use editor::{Editor, EditorElement, EditorStyle};
 use futures::{future::BoxFuture, stream::BoxStream, FutureExt, StreamExt};
 use gpui::{AnyView, AppContext, FontStyle, Task, TextStyle, View, WhiteSpace};
 use http::HttpClient;
-use open_ai::{stream_completion, Request, RequestMessage, Role as OpenAiRole};
+use language_model::Role;
 use settings::Settings;
 use std::time::Duration;
 use std::{env, sync::Arc};
@@ -18,103 +15,19 @@ use theme::ThemeSettings;
 use ui::prelude::*;
 use util::ResultExt;
 
-pub struct OpenAiCompletionProvider {
+pub struct AnthropicCompletionProvider {
     api_key: Option<String>,
     api_url: String,
-    model: OpenAiModel,
+    model: AnthropicModel,
     http_client: Arc<dyn HttpClient>,
     low_speed_timeout: Option<Duration>,
     settings_version: usize,
 }
 
-impl OpenAiCompletionProvider {
-    pub fn new(
-        model: OpenAiModel,
-        api_url: String,
-        http_client: Arc<dyn HttpClient>,
-        low_speed_timeout: Option<Duration>,
-        settings_version: usize,
-    ) -> Self {
-        Self {
-            api_key: None,
-            api_url,
-            model,
-            http_client,
-            low_speed_timeout,
-            settings_version,
-        }
-    }
-
-    pub fn update(
-        &mut self,
-        model: OpenAiModel,
-        api_url: String,
-        low_speed_timeout: Option<Duration>,
-        settings_version: usize,
-    ) {
-        self.model = model;
-        self.api_url = api_url;
-        self.low_speed_timeout = low_speed_timeout;
-        self.settings_version = settings_version;
-    }
-
-    fn to_open_ai_request(&self, request: LanguageModelRequest) -> Request {
-        let model = match request.model {
-            LanguageModel::OpenAi(model) => model,
-            _ => self.model.clone(),
-        };
-
-        Request {
-            model,
-            messages: request
-                .messages
-                .into_iter()
-                .map(|msg| match msg.role {
-                    Role::User => RequestMessage::User {
-                        content: msg.content,
-                    },
-                    Role::Assistant => RequestMessage::Assistant {
-                        content: Some(msg.content),
-                        tool_calls: Vec::new(),
-                    },
-                    Role::System => RequestMessage::System {
-                        content: msg.content,
-                    },
-                })
-                .collect(),
-            stream: true,
-            stop: request.stop,
-            temperature: request.temperature,
-            tools: Vec::new(),
-            tool_choice: None,
-        }
-    }
-}
-
-impl LanguageModelCompletionProvider for OpenAiCompletionProvider {
-    fn available_models(&self, cx: &AppContext) -> Vec<LanguageModel> {
-        if let AssistantProvider::OpenAi {
-            available_models, ..
-        } = &AssistantSettings::get_global(cx).provider
-        {
-            if !available_models.is_empty() {
-                return available_models
-                    .iter()
-                    .cloned()
-                    .map(LanguageModel::OpenAi)
-                    .collect();
-            }
-        }
-        let available_models = if matches!(self.model, OpenAiModel::Custom { .. }) {
-            vec![self.model.clone()]
-        } else {
-            OpenAiModel::iter()
-                .filter(|model| !matches!(model, OpenAiModel::Custom { .. }))
-                .collect()
-        };
-        available_models
-            .into_iter()
-            .map(LanguageModel::OpenAi)
+impl LanguageModelCompletionProvider for AnthropicCompletionProvider {
+    fn available_models(&self) -> Vec<LanguageModel> {
+        AnthropicModel::iter()
+            .map(LanguageModel::Anthropic)
             .collect()
     }
 
@@ -132,7 +45,7 @@ impl LanguageModelCompletionProvider for OpenAiCompletionProvider {
         } else {
             let api_url = self.api_url.clone();
             cx.spawn(|mut cx| async move {
-                let api_key = if let Ok(api_key) = env::var("OPENAI_API_KEY") {
+                let api_key = if let Ok(api_key) = env::var("ANTHROPIC_API_KEY") {
                     api_key
                 } else {
                     let (_, api_key) = cx
@@ -142,7 +55,7 @@ impl LanguageModelCompletionProvider for OpenAiCompletionProvider {
                     String::from_utf8(api_key)?
                 };
                 cx.update_global::<CompletionProvider, _>(|provider, _cx| {
-                    provider.update_current_as::<_, Self>(|provider| {
+                    provider.update_current_as::<_, AnthropicCompletionProvider>(|provider| {
                         provider.api_key = Some(api_key);
                     });
                 })
@@ -155,7 +68,7 @@ impl LanguageModelCompletionProvider for OpenAiCompletionProvider {
         cx.spawn(|mut cx| async move {
             delete_credentials.await.log_err();
             cx.update_global::<CompletionProvider, _>(|provider, _cx| {
-                provider.update_current_as::<_, Self>(|provider| {
+                provider.update_current_as::<_, AnthropicCompletionProvider>(|provider| {
                     provider.api_key = None;
                 });
             })
@@ -168,7 +81,7 @@ impl LanguageModelCompletionProvider for OpenAiCompletionProvider {
     }
 
     fn model(&self) -> LanguageModel {
-        LanguageModel::OpenAi(self.model.clone())
+        LanguageModel::Anthropic(self.model.clone())
     }
 
     fn count_tokens(
@@ -183,7 +96,7 @@ impl LanguageModelCompletionProvider for OpenAiCompletionProvider {
         &self,
         request: LanguageModelRequest,
     ) -> BoxFuture<'static, Result<BoxStream<'static, Result<String>>>> {
-        let request = self.to_open_ai_request(request);
+        let request = self.to_anthropic_request(request);
 
         let http_client = self.http_client.clone();
         let api_key = self.api_key.clone();
@@ -202,7 +115,19 @@ impl LanguageModelCompletionProvider for OpenAiCompletionProvider {
             let stream = response
                 .filter_map(|response| async move {
                     match response {
-                        Ok(mut response) => Some(Ok(response.choices.pop()?.delta.content?)),
+                        Ok(response) => match response {
+                            anthropic::ResponseEvent::ContentBlockStart {
+                                content_block, ..
+                            } => match content_block {
+                                anthropic::ContentBlock::Text { text } => Some(Ok(text)),
+                            },
+                            anthropic::ResponseEvent::ContentBlockDelta { delta, .. } => {
+                                match delta {
+                                    anthropic::TextDelta::TextDelta { text } => Some(Ok(text)),
+                                }
+                            }
+                            _ => None,
+                        },
                         Err(error) => Some(Err(error)),
                     }
                 })
@@ -217,50 +142,71 @@ impl LanguageModelCompletionProvider for OpenAiCompletionProvider {
     }
 }
 
-pub fn count_open_ai_tokens(
-    request: LanguageModelRequest,
-    background_executor: &gpui::BackgroundExecutor,
-) -> BoxFuture<'static, Result<usize>> {
-    background_executor
-        .spawn(async move {
-            let messages = request
+impl AnthropicCompletionProvider {
+    pub fn new(
+        model: AnthropicModel,
+        api_url: String,
+        http_client: Arc<dyn HttpClient>,
+        low_speed_timeout: Option<Duration>,
+        settings_version: usize,
+    ) -> Self {
+        Self {
+            api_key: None,
+            api_url,
+            model,
+            http_client,
+            low_speed_timeout,
+            settings_version,
+        }
+    }
+
+    pub fn update(
+        &mut self,
+        model: AnthropicModel,
+        api_url: String,
+        low_speed_timeout: Option<Duration>,
+        settings_version: usize,
+    ) {
+        self.model = model;
+        self.api_url = api_url;
+        self.low_speed_timeout = low_speed_timeout;
+        self.settings_version = settings_version;
+    }
+
+    fn to_anthropic_request(&self, mut request: LanguageModelRequest) -> Request {
+        request.preprocess_anthropic();
+
+        let model = match request.model {
+            LanguageModel::Anthropic(model) => model,
+            _ => self.model.clone(),
+        };
+
+        let mut system_message = String::new();
+        if request
+            .messages
+            .first()
+            .map_or(false, |message| message.role == Role::System)
+        {
+            system_message = request.messages.remove(0).content;
+        }
+
+        Request {
+            model,
+            messages: request
                 .messages
-                .into_iter()
-                .map(|message| tiktoken_rs::ChatCompletionRequestMessage {
-                    role: match message.role {
-                        Role::User => "user".into(),
-                        Role::Assistant => "assistant".into(),
-                        Role::System => "system".into(),
+                .iter()
+                .map(|msg| RequestMessage {
+                    role: match msg.role {
+                        Role::User => anthropic::Role::User,
+                        Role::Assistant => anthropic::Role::Assistant,
+                        Role::System => unreachable!("filtered out by preprocess_request"),
                     },
-                    content: Some(message.content),
-                    name: None,
-                    function_call: None,
+                    content: msg.content.clone(),
                 })
-                .collect::<Vec<_>>();
-
-            match request.model {
-                LanguageModel::Anthropic(_)
-                | LanguageModel::Cloud(CloudModel::Claude3_5Sonnet)
-                | LanguageModel::Cloud(CloudModel::Claude3Opus)
-                | LanguageModel::Cloud(CloudModel::Claude3Sonnet)
-                | LanguageModel::Cloud(CloudModel::Claude3Haiku)
-                | LanguageModel::OpenAi(OpenAiModel::Custom { .. }) => {
-                    // Tiktoken doesn't yet support these models, so we manually use the
-                    // same tokenizer as GPT-4.
-                    tiktoken_rs::num_tokens_from_messages("gpt-4", &messages)
-                }
-                _ => tiktoken_rs::num_tokens_from_messages(request.model.id(), &messages),
-            }
-        })
-        .boxed()
-}
-
-impl From<Role> for open_ai::Role {
-    fn from(val: Role) -> Self {
-        match val {
-            Role::User => OpenAiRole::User,
-            Role::Assistant => OpenAiRole::Assistant,
-            Role::System => OpenAiRole::System,
+                .collect(),
+            stream: true,
+            system: system_message,
+            max_tokens: 4092,
         }
     }
 }
@@ -295,7 +241,7 @@ impl AuthenticationPrompt {
         cx.spawn(|_, mut cx| async move {
             write_credentials.await?;
             cx.update_global::<CompletionProvider, _>(|provider, _cx| {
-                provider.update_current_as::<_, OpenAiCompletionProvider>(|provider| {
+                provider.update_current_as::<_, AnthropicCompletionProvider>(|provider| {
                     provider.api_key = Some(api_key);
                 });
             })
@@ -332,13 +278,11 @@ impl AuthenticationPrompt {
 
 impl Render for AuthenticationPrompt {
     fn render(&mut self, cx: &mut ViewContext<Self>) -> impl IntoElement {
-        const INSTRUCTIONS: [&str; 6] = [
-            "To use the assistant panel or inline assistant, you need to add your OpenAI API key.",
-            " - You can create an API key at: platform.openai.com/api-keys",
-            " - Make sure your OpenAI account has credits",
-            " - Having a subscription for another service like GitHub Copilot won't work.",
+        const INSTRUCTIONS: [&str; 4] = [
+            "To use the assistant panel or inline assistant, you need to add your Anthropic API key.",
+            "You can create an API key at: https://console.anthropic.com/settings/keys",
             "",
-            "Paste your OpenAI API key below and hit enter to use the assistant:",
+            "Paste your Anthropic API key below and hit enter to use the assistant:",
         ];
 
         v_flex()
@@ -360,7 +304,7 @@ impl Render for AuthenticationPrompt {
             )
             .child(
                 Label::new(
-                    "You can also assign the OPENAI_API_KEY environment variable and restart Zed.",
+                    "You can also assign the ANTHROPIC_API_KEY environment variable and restart Zed.",
                 )
                 .size(LabelSize::Small),
             )
