+++ crates/language_model/Cargo.toml
@@ -1,0 +1,41 @@
+[package]
+name = "language_model"
+version = "0.1.0"
+edition = "2021"
+publish = false
+license = "GPL-3.0-or-later"
+
+[lints]
+workspace = true
+
+[lib]
+path = "src/language_model.rs"
+doctest = false
+
+[features]
+test-support = [
+    "editor/test-support",
+    "language/test-support",
+    "project/test-support",
+    "text/test-support",
+]
+
+[dependencies]
+anthropic = { workspace = true, features = ["schemars"] }
+ollama = { workspace = true, features = ["schemars"] }
+open_ai = { workspace = true, features = ["schemars"] }
+schemars.workspace = true
+serde.workspace = true
+strum.workspace = true
+proto = { workspace = true, features = ["test-support"] }
+
+[dev-dependencies]
+ctor.workspace = true
+editor = { workspace = true, features = ["test-support"] }
+env_logger.workspace = true
+language = { workspace = true, features = ["test-support"] }
+log.workspace = true
+project = { workspace = true, features = ["test-support"] }
+rand.workspace = true
+text = { workspace = true, features = ["test-support"] }
+unindent.workspace = true
