+++ crates/semantic_index/Cargo.toml
@@ -22,6 +22,7 @@ anyhow.workspace = true
 client.workspace = true
 clock.workspace = true
 collections.workspace = true
+completion.workspace = true
 fs.workspace = true
 futures.workspace = true
 futures-batch.workspace = true
